requirements.txt
pylibmc==1.4.0
appnope==0.1.0
asn1crypto==0.23.0
backports-abc==0.5
bleach==2.0.0
certifi==2017.7.27.1
cffi==1.11.1
chardet==3.0.4
click==6.7
cloudinary==1.8.0
cryptography==2.0.3
decorator==4.0.11
dj-database-url==0.2.1
dj-static==0.0.6
Django==1.9.7
django-appconf==1.0.2
django-core==1.4.1
django-debug-toolbar==1.8
django-extensions==1.7.9
django-forms-bootstrap==3.1.0
django-jsonfield==1.0.1
django-paypal==0.4.1
django-registration-redux==1.7
django-storages==1.6.4
django-stripe-payments==2.0.0
django-tabination==0.4.0
django-toolbelt==0.0.1
entrypoints==0.2.3
Flask==0.12.2
gunicorn==19.6.0
html5lib==0.999999999
idna==2.6
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
jsonfield==1.0.3
jsonschema==2.6.0
jupyter-client==5.0.1
jupyter-core==4.3.0
MarkupSafe==1.0
mistune==0.7.4
mock==2.0.0
nbconvert==5.2.1
nbformat==4.3.0
ndg-httpsclient==0.4.3
notebook==5.0.0
olefile==0.44
pandocfilters==1.4.1
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
pinax-stripe==3.4.1
prompt-toolkit==1.0.14
psycopg2==2.6.2
ptyprocess==0.5.1
pyasn1==0.3.7
pycparser==2.18
Pygments==2.2.0
pyOpenSSL==17.3.0
python-dateutil==2.6.0
python-http-client==2.2.1
pytz==2017.2
pyzmq==16.0.2
requests==2.18.4
requirements==0.1
sendgrid==4.2.1
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.3
static3==0.7.0
stripe==1.19.1
terminado==0.6
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
typing==3.6.1
uploader==0.4.1
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
wheel==0.29.0
whitenoise==2.0.6
widgetsnbextension==2.0.0
brotlipy==0.7.0
django-filters==0.2.1
